tee-time-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tee-time-agent"
version = "0.1.0"
description = "Automation agent that uses Playwright and Google AI to collect BRS tee sheet availability and deliver updates to Telegram."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "playwright>=1.45.0",
  "google-adk>=1.16.0",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "python-dateutil>=2.9.0",
  "structlog>=24.1.0",
  "tenacity>=8.3.0"
]

[project.scripts]
tee-time-agent = "tee_time_agent.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
